Parser for reading the data file
This script creates a dictionary/ map of nodes (where Node is a class type of 2 strings: name, color) and an ArrayList of edges, where an edge is a class type containing 2 nodes. 
I have not implemented the building of the graph, as this might differ from algorithm to algorithm. However, one way to implement a graph is to simply create a dictionary: "graph", run through the ArrayList of edges, and map each edge, to its nodes.

The script ends with 3 lines of printing, this is for debugging purposes only and may be removed, once you have convinced yourself that it passes the input file correctly.