My solution to problem: None
Creates a graph (dictionary), mapping a valid node to all its edges. A node is valid if it is a sink or source node (regardless of color) or a black node.  Breadth-first search is then run on the graph and a path is calculated. The length of the path is returned, unless there exists no path from the source to the sink node, in which case -1 is returned. 
Running this code on the G-ex.txt returns the expected value.